backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
httpx==0.27.0
openai==1.35.3
scikit-learn==1.5.0
numpy==1.26.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
pytest==8.2.2
pytest-asyncio==0.23.7
